view controller to view model
