--- firebaseRealtimeDBExample/firebaseRealtimeDBExample/Scenes/Basket/BasketViewModel.swift
@@ -1,8 +1,0 @@
-//
-//  BasketViewModel.swift
-//  firebaseRealtimeDBExample
-//
-//  Created by Dilda Ezgi Metincan on 14.02.2023.
-//
-
-import Foundation
