--- firebaseRealtimeDBExample/firebaseRealtimeDBExample/Scenes/ProductDetail/ProductDetailViewModel.swift
@@ -1,8 +1,0 @@
-//
-//  ProductDetailViewModel.swift
-//  firebaseRealtimeDBExample
-//
-//  Created by Dilda Ezgi Metincan on 14.02.2023.
-//
-
-import Foundation
