+++ firebaseRealtimeDBExample/firebaseRealtimeDBExample/Scenes/ProductList/ProductListVC.swift
@@ -10,20 +10,19 @@ import Firebase
 import UIKit
 
 class ProductListVC: UIViewController, UISearchBarDelegate {
-    let database = Database.database().reference()
-    let databaseRef = Database.database().reference(withPath: "products")
     let navigator = Navigator()
-
     var collectionView: UICollectionView!
-    var data: [Product] = []
-    var filteredItems: [Product] = []
+
     
     private let spacing : CGFloat = 15.0
+    private let viewModel = ProductListViewModel()
 
     override func viewDidLoad() {
         super.viewDidLoad()
-        collectionViewDesign()
-        fetchObjects()
+        setupView()
+        viewModel.fetchObjects() {
+            self.collectionView.reloadData()
+        }
         navigator.navController = navigationController!
         let basketButton = UIButton(type: .system)
         basketButton.imageView?.image = UIImage(named: "basketImage")
@@ -37,28 +36,20 @@ class ProductListVC: UIViewController, UISearchBarDelegate {
     
     func scrollViewDidScroll(_ scrollView: UIScrollView) {
         if scrollView.contentOffset.y > 0 {
-                        navigationController?.navigationBar.isHidden = false
-                        navigationController?.navigationBar.backgroundColor = .white
-                        navigationController?.navigationBar.isOpaque = false
+            navigationController?.navigationBar.isHidden = false
+            navigationController?.navigationBar.backgroundColor = .white
+            navigationController?.navigationBar.isOpaque = false
         }
-
+        
     }
     
     @objc func addToBasketButtonTapped(_ sender: UIButton) {
-        print("basildi")
-        
         if let indexPath = collectionView.indexPath(for: sender.superview?.superview as! UICollectionViewCell) {
-            let product = data[indexPath.row]
-            
+            let product = viewModel.data[indexPath.row]
             var addedProducts = UserDefaults.standard.array(forKey: "basket") as? [[String: Any]] ?? []
-            
-            // Yeni ürünü sepete ekle
             let newProduct = ["productID": product.productID, "name": product.productName, "price": product.productPrice, "rate": product.productRate, "image": product.productImages.image1 ?? ""] as [String : Any]
             addedProducts.append(newProduct)
-            
-            // Sepete eklenen ürünleri güncelle
             UserDefaults.standard.set(addedProducts, forKey: "basket")
-            
             let basketVC = BasketVC()
             navigator.navigateTo(basketVC, animated: true)
         }
@@ -69,7 +60,7 @@ class ProductListVC: UIViewController, UISearchBarDelegate {
         navigator.navigateTo(basketVC, animated: true)
     }
     
-    func collectionViewDesign() {
+    func setupView() {
         let layout = UICollectionViewFlowLayout()
         layout.scrollDirection = .vertical
         layout.sectionInset = UIEdgeInsets(top: spacing + 75, left: spacing, bottom: spacing, right: spacing)
@@ -111,42 +102,18 @@ class ProductListVC: UIViewController, UISearchBarDelegate {
         searchBar.heightAnchor.constraint(equalToConstant: 44).isActive = true 
     }
     
-    func fetchObjects() {
-        // get data from Firebase
-        databaseRef.observeSingleEvent(of: .value) { snapshot in
-            guard let dict = snapshot.value as? [String: Any] else {
-                return
-            }
-            for (_, value) in dict {
-                if let productDict = value as? [String: Any] {
-                    let productName = productDict["productName"] as? String
-                    let productPrice = productDict["productPrice"] as? Double
-                    let productRate = productDict["productRate"] as? Float
-                    let productImagesDict = productDict["productImages"] as? [String: String]
-                    let productID = productDict["productID"] as? String
-                    
-                    let productImages = ProductImages(image1: productImagesDict?["image1"] ?? "", image2: productImagesDict?["image2"] ?? "")
-                    let product = Product(productName: productName ?? "", productID: productID ?? "", productPrice: productPrice ?? 0.0, productRate: productRate ?? 0.0, productImages: productImages)
-                    self.data.append(product)
-                }
-            }
-            self.filteredItems = self.data // Filtrelenmiş öğeler için data dizisi
-            self.collectionView.reloadData()
-        }
-    }
-    
     func searchBar(_ searchBar: UISearchBar, textDidChange searchText: String) {
-        filteredItems = []
+        viewModel.filteredItems = []
         if searchText == "" {
-            filteredItems = data // Arama yapılmadığında tüm öğeleri gösterme
+            viewModel.filteredItems = viewModel.data
         } else {
-            for item in data {
+            for item in viewModel.data {
                 if item.productName.lowercased().contains(searchText.lowercased()) {
-                    filteredItems.append(item)
+                    viewModel.filteredItems.append(item)
                 }
             }
         }
-        collectionView.reloadData() // Filtrelenmiş öğeleri görüntüle
+        collectionView.reloadData()
     }
     
     func searchBarShouldBeginEditing(_ searchBar: UISearchBar) -> Bool {
@@ -157,7 +124,7 @@ class ProductListVC: UIViewController, UISearchBarDelegate {
 
 extension ProductListVC: UICollectionViewDelegateFlowLayout, UICollectionViewDataSource {
     func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return filteredItems.count
+        return viewModel.filteredItems.count
     }
 
     func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
@@ -168,7 +135,7 @@ extension ProductListVC: UICollectionViewDelegateFlowLayout, UICollectionViewDat
         cell.layer.borderColor = UIColor(named: "borderColor")?.cgColor
         cell.addToBasketButton.addTarget(self, action: #selector(addToBasketButtonTapped(_:)), for: .touchUpInside)
         
-        let currentData = filteredItems[indexPath.row]
+        let currentData = viewModel.filteredItems[indexPath.row]
         cell.configure(withData: currentData)
         
         return cell
@@ -185,7 +152,7 @@ extension ProductListVC: UICollectionViewDelegateFlowLayout, UICollectionViewDat
     }
     
     func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
-        let product = data[indexPath.item]
+        let product = viewModel.data[indexPath.item]
         let productDetailVC = ProductDetailVC(product: product)
         navigator.navigateTo(productDetailVC, animated: true)
     }
